requirements.txt
aiogram==3.15.0
gigachat==0.1.37
asyncpg==0.30.0
python-dotenv==1.0.1
orjson==3.10.12
loguru==0.7.2
pydantic
pydantic-settings
